+++ spring-skillary-back/src/main/java/com/example/springskillaryback/controller/ContentController.java
@@ -4,8 +4,10 @@ import com.example.springskillaryback.common.dto.CategoryResponseDto;
 import com.example.springskillaryback.common.dto.ContentListResponseDto;
 import com.example.springskillaryback.common.dto.ContentRequestDto;
 import com.example.springskillaryback.common.dto.ContentResponseDto;
+import com.example.springskillaryback.common.util.TokenUtil;
 import com.example.springskillaryback.domain.CategoryEnum;
 import com.example.springskillaryback.service.ContentService;
+import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Slice;
 import org.springframework.http.HttpStatus;
@@ -20,14 +22,17 @@ import java.util.List;
 @RequiredArgsConstructor
 public class ContentController {
 	private final ContentService contentService;
+	private final TokenUtil tokenUtil;
 
 	@PostMapping
 	public ResponseEntity<ContentResponseDto> createContent(
 		@RequestBody ContentRequestDto requestDto,
-		@RequestHeader("X-Creator-Id") Byte creatorId // [임시] 시큐리티 작업전
-        // @AuthenticationPrincipal CustomPrincipal customPrincipal
+		HttpServletRequest request
 	) {
-        // Byte creatorId = customPrincipal.getCreatorId();
+		Byte creatorId = tokenUtil.getCreatorIdFromToken(request);
+		if (creatorId == null) {
+			return ResponseEntity.status(HttpStatus.FORBIDDEN).build(); // 403 크리에이터만 가능
+		}
 		ContentResponseDto response = contentService.createContent(requestDto, creatorId);
 		return ResponseEntity.status(HttpStatus.CREATED).body(response); // 201
 	}
@@ -36,10 +41,12 @@ public class ContentController {
 	public ResponseEntity<ContentResponseDto> updateContent(
 		@PathVariable Byte contentId,
 		@RequestBody ContentRequestDto requestDto,
-		@RequestHeader("X-Creator-Id") Byte creatorId // [임시] 시큐리티 작업전
-        // @AuthenticationPrincipal CustomPrincipal customPrincipal
+		HttpServletRequest request
 	) {
-        // Byte creatorId = customPrincipal.getCreatorId();
+		Byte creatorId = tokenUtil.getCreatorIdFromToken(request);
+		if (creatorId == null) {
+			return ResponseEntity.status(HttpStatus.FORBIDDEN).build(); // 403 크리에이터만 가능
+		}
 		ContentResponseDto response = contentService.updateContent(contentId, requestDto, creatorId);
 		return ResponseEntity.ok(response); // 200
 	}
@@ -105,10 +112,12 @@ public class ContentController {
 	@DeleteMapping("/{contentId}")
 	public ResponseEntity<Void> deleteContent(
 		@PathVariable Byte contentId,
-		@RequestHeader("X-Creator-Id") Byte creatorId // [임시] 시큐리티 작업전
-        // @AuthenticationPrincipal CustomPrincipal customPrincipal
+		HttpServletRequest request
 	) {
-        // Byte creatorId = customPrincipal.getCreatorId();
+		Byte creatorId = tokenUtil.getCreatorIdFromToken(request);
+		if (creatorId == null) {
+			return ResponseEntity.status(HttpStatus.FORBIDDEN).build(); // 403 크리에이터만 가능
+		}
 
 		// Content 삭제 (파일 삭제까지 포함하여 처리)
 		contentService.deleteContent(contentId, creatorId);
